Tried to rationalize default constructor of some fields as per my comment in issue #203 Quick question about implementation. This was implemented as:
    public MCheckBox() {
    }
Shouldn't it be:

    public MCheckBox() {
        super();
    }
Small distinction but if you use the other constructors, they all call super (e.g. super(caption)) and the super constructor generally calls this() which in the case of CheckBox contains:

    public CheckBox() {
        registerRpc(rpc);
        registerRpc(focusBlurRpc);
        setValue(Boolean.FALSE);
    }
So if you use the MCheckBox constructor with the caption, or with the caption and an initial state, that registerRpc stuff gets called and if you use the empty constructor it does not. For consistency sake it should probably all follow a common path.

The same is probably true for MDateField, MTextField, MTextArea, EmailField although in those cases it makes less difference.